# krishi_connect/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Farmer Platform API",
        "description": "API for farmers in Maharashtra with market rates, schemes, and AI advisory.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/status": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Status",
                "operationId": "admin_status_api_v1_admin_status_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/admin/crops": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create New Crop",
                "operationId": "create_new_crop_api_v1_admin_crops_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CropCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Crop"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/admin/crops/{crop_id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update Existing Crop",
                "operationId": "update_existing_crop_api_v1_admin_crops__crop_id__put",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "crop_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Crop Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CropUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Crop"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete Crop",
                "operationId": "delete_crop_api_v1_admin_crops__crop_id__delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "crop_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Crop Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/schemes": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create New Scheme",
                "operationId": "create_new_scheme_api_v1_admin_schemes_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GovernmentSchemeCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GovernmentScheme"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/admin/schemes/{scheme_id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update Existing Scheme",
                "operationId": "update_existing_scheme_api_v1_admin_schemes__scheme_id__put",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "scheme_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Scheme Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GovernmentSchemeUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GovernmentScheme"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete Scheme",
                "operationId": "delete_scheme_api_v1_admin_schemes__scheme_id__delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "scheme_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Scheme Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/market_rates": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create New Market Rate",
                "operationId": "create_new_market_rate_api_v1_admin_market_rates_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MarketRateCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarketRate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/admin/market_rates/{rate_id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update Existing Market Rate",
                "operationId": "update_existing_market_rate_api_v1_admin_market_rates__rate_id__put",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "rate_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Rate Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MarketRateUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarketRate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete Market Rate",
                "operationId": "delete_market_rate_api_v1_admin_market_rates__rate_id__delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "rate_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Rate Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Crop": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Name"
                    },
                    "image_url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 512
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Url"
                    },
                    "details": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Details"
                    },
                    "context": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Context"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "id"
                ],
                "title": "Crop"
            },
            "CropCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Name"
                    },
                    "image_url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 512
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Url"
                    },
                    "details": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Details"
                    },
                    "context": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "CropCreate"
            },
            "CropUpdate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Name"
                    },
                    "image_url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 512
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Url"
                    },
                    "details": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Details"
                    },
                    "context": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "CropUpdate"
            },
            "GovernmentScheme": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 512,
                        "title": "Title"
                    },
                    "link": {
                        "type": "string",
                        "maxLength": 512,
                        "title": "Link"
                    },
                    "image_url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 512
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Url"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "region": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Region",
                        "default": "Maharashtra"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "created_by_id": {
                        "type": "integer",
                        "title": "Created By Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "link",
                    "id",
                    "created_by_id",
                    "created_at"
                ],
                "title": "GovernmentScheme"
            },
            "GovernmentSchemeCreate": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 512,
                        "title": "Title"
                    },
                    "link": {
                        "type": "string",
                        "maxLength": 512,
                        "title": "Link"
                    },
                    "image_url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 512
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Url"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "region": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Region",
                        "default": "Maharashtra"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "link"
                ],
                "title": "GovernmentSchemeCreate"
            },
            "GovernmentSchemeUpdate": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 512,
                        "title": "Title"
                    },
                    "link": {
                        "type": "string",
                        "maxLength": 512,
                        "title": "Link"
                    },
                    "image_url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 512
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Url"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "region": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Region",
                        "default": "Maharashtra"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "link"
                ],
                "title": "GovernmentSchemeUpdate"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "MarketRateCreate": {
                "properties": {
                    "crop_id": {
                        "type": "integer",
                        "title": "Crop Id"
                    },
                    "location": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Location"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "title": "Date"
                    },
                    "min_price": {
                        "type": "number",
                        "exclusiveMinimum": 0.0,
                        "title": "Min Price"
                    },
                    "max_price": {
                        "type": "number",
                        "exclusiveMinimum": 0.0,
                        "title": "Max Price"
                    },
                    "expected_price": {
                        "anyOf": [
                            {
                                "type": "number",
                                "exclusiveMinimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Expected Price"
                    },
                    "unit": {
                        "type": "string",
                        "maxLength": 50,
                        "title": "Unit",
                        "default": "Quintal"
                    }
                },
                "type": "object",
                "required": [
                    "crop_id",
                    "location",
                    "date",
                    "min_price",
                    "max_price"
                ],
                "title": "MarketRateCreate"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login"
                    }
                }
            }
        }
    }
}